@@ -1,88 +1,118 @@
 {
-  "name": "projection-path-explorer",
+  "name": "projection-space-explorer",
+  "description": "",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "dist/bundle.js",
+  "types": "dist/index.d.ts",
   "author": "Moritz Schöfl",
-  "license": "MIT",
+  "license": "BSD 3-Clause License",
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.2",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@types/concaveman": "^1.1.3",
-    "@types/d3": "^5.7.2",
+    "@material-ui/core": "^4.12.3",
+    "@material-ui/icons": "^4.11.2",
+    "@mui/icons-material": "^5.0.5",
+    "@mui/lab": "^5.0.0-alpha.53",
+    "@reduxjs/toolkit": "^1.6.2",
+    "@tootallnate/once": "^3.0.0",
+    "@types/d3v5": "npm:@types/d3@^5.7.2",
     "@types/jest": "^26.0.15",
-    "@types/node": "^14.0.11",
-    "@types/react-redux": "^7.1.9",
+    "@types/node": "^14.17.32",
+    "@types/offscreencanvas": "^2019.7.0",
+    "@types/papaparse": "^5.3.1",
+    "@types/react": "^17.0.34",
+    "@types/react-dom": "^16.9.5",
+    "@types/react-redux": "^7.1.20",
+    "cat-rom-spline": "^1.0.0",
     "clone": "^2.1.2",
-    "concaveman": "^1.1.1",
-    "conrec": "^1.0.0",
-    "core-js": "^3.6.4",
-    "d3": "^5.16.0",
+    "colorette": "^2.0.16",
     "d3-contour": "^2.0.0",
+    "d3v5": "npm:d3@5.16.0",
     "fast-clone": "^1.5.13",
     "graphology": "^0.17.1",
     "graphology-layout-forceatlas2": "^0.4.4",
-    "graphology-shortest-path": "^1.3.4",
+    "graphology-shortest-path": "^1.4.2",
+    "graphology-types": "^0.21.2",
+    "install": "^0.13.0",
     "jaccard": "^0.4.0",
     "jsfive": "^0.2.1",
     "jstat": "^1.9.4",
-    "libtess": "^1.2.2",
-    "lineupjs": "^4.1.2",
-    "lineupjsx": "^4.0.0",
     "lodash": "^4.17.20",
-    "ml-conrec": "^3.2.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-draggable": "^4.4.3",
-    "react-icons": "^4.1.0",
-    "react-loader-spinner": "^3.1.14",
+    "papaparse": "^5.3.1",
+    "react-data-grid": "7.0.0-beta.16",
     "react-promise-tracker": "^2.1.0",
-    "react-redux": "^7.2.0",
     "react-rnd": "^10.2.2",
     "react-split": "^2.0.9",
-    "react-vega-lite": "^6.1.0",
-    "react-virtualized": "^9.22.2",
-    "redux": "^4.0.5",
+    "react-vega": "^7.4.4",
     "redux-devtools-extension": "^2.13.8",
+    "redux-thunk": "^2.4.0",
     "regenerator-runtime": "^0.13.3",
     "resize-observer": "^1.0.0",
+    "sass": "^1.55.0",
     "serve-dist": "^1.0.11",
     "sigma-webpack": "^1.2.0-1",
     "three": "^0.117.1",
     "three.meshline": "^1.2.0",
     "tinyqueue": "^2.0.3",
     "umap-js": "^1.3.2",
-    "vega": "^5.17.0",
-    "vega-lite": "^4.16.8",
-    "vega-tooltip": "^0.24.2"
+    "url-loader": "^4.1.1",
+    "vega": "5.17.0",
+    "vega-lite": "4.16.8",
+    "vega-typings": "0.19.1"
   },
   "devDependencies": {
     "@babel/core": "^7.10.2",
     "@babel/preset-env": "^7.10.2",
     "@babel/preset-react": "^7.8.3",
     "@babel/preset-typescript": "^7.8.3",
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
+    "@mui/material": "^5.0.6",
+    "@mui/styles": "^5.0.2",
     "@svgr/webpack": "^5.5.0",
-    "@types/react": "^16.9.19",
-    "@types/react-dom": "^16.9.5",
-    "babel-loader": "^8.0.6",
+    "@typescript-eslint/eslint-plugin": "^5.9.0",
+    "@typescript-eslint/parser": "^5.9.0",
+    "babel-loader": "^8.2.3",
     "css-loader": "^3.4.2",
+    "eslint": "^8.7.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^16.1.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jest": "^25.3.4",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "jest": "^26.6.3",
-    "node-sass": "^4.13.1",
+    "prettier": "^2.5.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.6",
+    "redux": "^4.1.2",
     "sass-loader": "^8.0.2",
-    "shader-loader": "^1.3.1",
     "source-map-loader": "^0.2.4",
     "style-loader": "^1.1.3",
     "svg-url-loader": "^7.1.1",
-    "ts-loader": "^6.2.1",
-    "typescript": "^3.7.5",
-    "webpack": "^4.41.5",
-    "webpack-cli": "^3.3.10",
-    "webpack-glsl-loader": "^1.0.1"
+    "ts-loader": "^8.2.0",
+    "typescript": "^4.7.4",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0",
+    "worker-loader": "^3.0.8"
   },
+  "peerDependencies": {
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
+    "@mui/material": "^5.0.3",
+    "@mui/styles": "^5.0.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.0",
+    "redux": "^4.0.5"
+  },
+  
   "scripts": {
-    "webpack:dev": "npx webpack --config webpack.config.js",
-    "test": "jest"
+    "build": "npx webpack --config webpack.config.js --stats-children",
+    "test": "jest",
+    "lint": "eslint src/**/*.ts{,x}",
+    "lint:fix": "eslint --fix src/**/*.ts{,x}"
   }
 }
