@@ -1,57 +1,67 @@
-import { Grid } from "@material-ui/core";
-import React = require("react");
-import useCancellablePromise from "../../../utils/promise-helpers";
-import { CSVLoader } from "../../Utility/Loaders/CSVLoader";
-import { JSONLoader } from "../../Utility/Loaders/JSONLoader";
-import { SDFLoader } from "../../Utility/Loaders/SDFLoader";
-import DragAndDrop from "./DragAndDrop";
-import { SDFModifierDialog } from "./SDFModifierDialog";
+import { Grid } from '@mui/material';
+import * as React from 'react';
+import * as Papa from 'papaparse';
+import type { Dataset } from '../../..';
+import { CSVLoader } from '../../Utility/Loaders/CSVLoader';
+import { JSONLoader } from '../../Utility/Loaders/JSONLoader';
+import { DragAndDrop } from './DragAndDrop';
 
-export var DatasetDrop = ({ onChange, cancellablePromise, abort_controller }) => {
-    const [entry, setEntry] = React.useState(null);
-    const [openSDFDialog, setOpen] = React.useState(false);
-    
+export function DatasetDrop({ onChange }: { onChange(dataset: Dataset): void }) {
+  return (
+    <Grid container item alignItems="stretch" justifyContent="center" direction="column" style={{ padding: '16px' }}>
+      <DragAndDrop
+        accept=".csv,.json,.sdf"
+        handleDrop={(files) => {
+          if (files == null || files.length <= 0) {
+            return;
+          }
 
-    function onModifierDialogClose(modifiers){
-        setOpen(false); 
-        if(modifiers !== null){
-            abort_controller = new AbortController();
-            new SDFLoader().resolveContent(entry, onChange, cancellablePromise, modifiers, abort_controller);
-        }
-    }
+          const file = files[0];
+          const fileName = file.name as string;
 
-    return <Grid container item alignItems="stretch" justify="center" direction="column" style={{ padding: '16px' }}>
-        <DragAndDrop accept="image/*" handleDrop={(files) => {
-            if (files == null || files.length <= 0) {
-                return;
-            }
+          if (fileName.endsWith('.csv')) {
+            const rows = [];
+            let header = null;
 
-            var file = files[0]
-            var fileName = file.name as string
+            Papa.parse(file, {
+              // @ts-ignore
+              worker: true,
+              skipEmptyLines: true,
+              step(results: any) {
+                if (!header) {
+                  header = results.data;
 
-            if(fileName.endsWith('sdf')){
-                setEntry(file);
-                setOpen(true);
-            }else{
-
-                var reader = new FileReader()
-                reader.onload = (event) => {
-                    var content = event.target.result
-
-                    if (fileName.endsWith('json')) {
-                        new JSONLoader().resolveContent(content, onChange)
-                    } else {
-                        new CSVLoader().resolveContent(content, onChange)
-                    }
+                  return;
                 }
 
-                reader.readAsText(file)
-            }
+                const dict = {};
+                results.data.forEach((value, i) => {
+                  dict[header[i]] = value;
+                });
 
+                rows.push(dict);
+              },
+              complete() {
+                new CSVLoader().resolveVectors(rows, onChange);
+              },
+            });
+          } else {
+            const reader = new FileReader();
 
-        }}>
-            <div style={{ height: 200 }}></div>
-        </DragAndDrop>
-        <SDFModifierDialog openSDFDialog={openSDFDialog} handleClose={onModifierDialogClose}></SDFModifierDialog>
+            reader.onload = (event) => {
+              const content = event.target.result;
+
+              if (fileName.endsWith('json')) {
+                new JSONLoader().resolveContent(content, onChange);
+              }
+            };
+
+            reader.readAsText(file);
+          }
+        }}
+      >
+        <div style={{ height: 200 }} />
+      </DragAndDrop>
     </Grid>
+  );
 }
