@@ -1,280 +1,365 @@
-import { connect, ConnectedProps } from 'react-redux'
-import React = require('react')
-import { FlexParent } from '../../Utility/FlexParent'
-import { Avatar, Box, Button, Grid, IconButton, List, ListItem, ListItemAvatar, ListItemSecondaryAction, ListItemText, Typography } from '@material-ui/core'
-import { ProjectionControlCard } from './ProjectionControlCard/ProjectionControlCard'
-import { setProjectionOpenAction } from "../../Ducks/ProjectionOpenDuck"
-import { setProjectionWorkerAction } from "../../Ducks/ProjectionWorkerDuck"
-import { Dataset } from "../../Utility/Data/Dataset"
-import { GenericSettings } from './GenericSettings/GenericSettings'
-import { RootState } from '../../Store/Store'
-import { setProjectionParamsAction } from '../../Ducks/ProjectionParamsDuck'
-import { setProjectionColumns } from '../../Ducks/ProjectionColumnsDuck'
-import { TSNEEmbeddingController } from './EmbeddingController/TSNEEmbeddingController'
-import { UMAPEmbeddingController } from './EmbeddingController/UMAPEmbeddingController'
-import { ClusterTrailSettings } from './ClusterTrailSettings/ClusterTrailSettings'
-import { setTrailVisibility } from '../../Ducks/TrailSettingsDuck'
-import { ForceAtlas2EmbeddingController } from './EmbeddingController/ForceAtlas2EmbeddingController'
-import { Embedding } from '../../Utility/Data/Embedding'
-import FolderIcon from '@material-ui/icons/Folder';
-import DeleteIcon from '@material-ui/icons/Delete';
-import { addProjectionAction, deleteProjectionAction } from '../../Ducks/ProjectionsDuck'
-
-import * as frontend_utils from '../../../utils/frontend-connect';
+import { connect, ConnectedProps, useDispatch, useSelector } from 'react-redux';
+import * as React from 'react';
+import { Box, Button, Chip, Grid, IconButton, List, ListItem, ListItemSecondaryAction, ListItemText, Typography } from '@mui/material';
+import SettingsIcon from '@mui/icons-material/Settings';
+import { EntityId } from '@reduxjs/toolkit';
+import { ProjectionControlCard } from './ProjectionControlCard';
+import { setProjectionOpenAction } from '../../Ducks/ProjectionOpenDuck';
+import { setProjectionWorkerAction } from '../../Ducks/ProjectionWorkerDuck';
+import { ADataset, Dataset } from '../../../model/Dataset';
+import { GenericSettings } from './GenericSettings';
+import type { RootState } from '../../Store/Store';
+import { TSNEEmbeddingController } from './TSNEEmbeddingController';
+import { UMAPEmbeddingController } from './UMAPEmbeddingController';
+import { ClusterTrailSettings } from './ClusterTrailSettings';
+import { setTrailVisibility } from '../../Ducks/TrailSettingsDuck';
+import { ForceAtlas2EmbeddingController } from './ForceAtlas2EmbeddingController';
+import { IProjection, IBaseProjection, ProjectionMethod, IPosition } from '../../../model/ProjectionInterfaces';
+import { FeatureConfig, DEFAULT_EMBEDDINGS, EmbeddingMethod } from '../../../BaseConfig';
+import { EditProjectionDialog } from './EditProjectionDialog';
+import { ViewActions, ViewSelector } from '../../Ducks/ViewDuck';
+import { FeatureType } from '../../../model';
+import { SelectFeatureComponent } from '../StatesTabPanel/SelectFeatureComponent';
+import { setProjectionColumnsAction } from '../../Ducks';
 
 const mapStateToProps = (state: RootState) => ({
-    currentAggregation: state.currentAggregation,
-    stories: state.stories,
-    projectionWorker: state.projectionWorker,
-    projectionOpen: state.projectionOpen,
-    dataset: state.dataset,
-    webGLView: state.webGLView,
-    projectionParams: state.projectionParams,
-    projections: state.projections
-})
+  // currentAggregation: state.currentAggregation,
+  stories: state.stories,
+  projectionWorker: state.projectionWorker,
+  projectionOpen: state.projectionOpen,
+  dataset: state.dataset,
+  // projectionParams: state.projectionParams,
+  projections: state.multiples.projections,
+  projectionParams: state.projectionParams,
+});
 
-const mapDispatchToProps = dispatch => ({
-    setProjectionOpen: value => dispatch(setProjectionOpenAction(value)),
-    setProjectionWorker: value => dispatch(setProjectionWorkerAction(value)),
-    setProjectionParams: value => dispatch(setProjectionParamsAction(value)),
-    setProjectionColumns: value => dispatch(setProjectionColumns(value)),
-    setTrailVisibility: visibility => dispatch(setTrailVisibility(visibility)),
-    addProjection: embedding => dispatch(addProjectionAction(embedding)),
-    deleteProjection: projection => dispatch(deleteProjectionAction(projection))
-})
+const mapDispatchToProps = (dispatch) => ({
+  setProjectionOpen: (value) => dispatch(setProjectionOpenAction(value)),
+  setProjectionWorker: (value) => dispatch(setProjectionWorkerAction(value)),
+  // setProjectionParams: value => dispatch(setProjectionParamsAction(value)),
+  setProjectionColumns: (value) => dispatch(setProjectionColumnsAction(value)),
+  setTrailVisibility: (visibility) => dispatch(setTrailVisibility(visibility)),
+  addProjection: (embedding) => dispatch(ViewActions.add(embedding)),
+  deleteProjection: (handle: string) => dispatch(ViewActions.remove(handle)),
+  updateWorkspace: (workspace: IBaseProjection, metadata?) => dispatch(ViewActions.updateActive({ positions: workspace, metadata })),
+});
 
 const connector = connect(mapStateToProps, mapDispatchToProps);
 
-type PropsFromRedux = ConnectedProps<typeof connector>
+type PropsFromRedux = ConnectedProps<typeof connector>;
 
 type Props = PropsFromRedux & {
-    projectionWorker?: Worker
-    projectionOpen?: boolean
-    setProjectionOpen?: any
-    setProjectionWorker?: any
-    dataset?: Dataset
-    webGLView?: any
+  config: FeatureConfig;
+  projectionWorker?: Worker;
+  projectionOpen?: boolean;
+  setProjectionOpen?: any;
+  setProjectionWorker?: any;
+  dataset?: Dataset;
+};
+
+function EmbeddingMethodButtons(props: { setOpen; setDomainSettings; embeddings?: EmbeddingMethod[] }) {
+  const embeddings = props.embeddings ?? DEFAULT_EMBEDDINGS;
+
+  return (
+    <Grid container direction="column" spacing={1}>
+      {embeddings.map((emb) => (
+        <Grid key={emb.id} item>
+          <Button
+            data-cy={`embedding-${emb.id}`}
+            style={{
+              width: '100%',
+            }}
+            variant="outlined"
+            onClick={() => {
+              props.setDomainSettings(emb);
+              props.setOpen(true);
+            }}
+          >
+            {emb.name}
+          </Button>
+        </Grid>
+      ))}
+    </Grid>
+  );
 }
 
-
 export const EmbeddingTabPanel = connector((props: Props) => {
-    const [open, setOpen] = React.useState(false)
-    const [domainSettings, setDomainSettings] = React.useState('')
+  const [open, setOpen] = React.useState(false);
+  const [domainSettings, setDomainSettings] = React.useState({
+    id: '',
+    name: '',
+    embController: null,
+  });
+  const [controller, setController] = React.useState(null);
+  const [projectionToEdit, setProjectionToEdit] = React.useState<IProjection>(null);
+  const [step, setStep] = React.useState(0);
+  const stepRef = React.useRef(0);
 
-    const [controller, setController] = React.useState(null)
+  const dispatch = useDispatch();
 
-    React.useEffect(()=>{
-        if (controller) {
-            controller.terminate()
-        }
-        setController(null)
-        props.setTrailVisibility(false)
-    }, [props.dataset]);
+  const workspace = useSelector(ViewSelector.getWorkspace);
+  const workspaceIsTemporal = useSelector(ViewSelector.workspaceIsTemporal);
+  const dataset = useSelector((state: RootState) => state.dataset);
 
+  let numericFeatures = null;
+  if (dataset) {
+    numericFeatures = Object.keys(dataset.columns).filter((key) => dataset.columns[key].featureType === FeatureType.Quantitative);
+  }
 
-    const onSaveProjectionClick = (name) => {
-        props.addProjection(new Embedding(props.dataset.vectors, "Created " + name))
+  React.useEffect(() => {
+    if (controller) {
+      controller.terminate();
     }
+    setController(null);
+    props.setTrailVisibility(false);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [props.dataset]);
 
-    const onProjectionClick = (projection: Embedding) => {
-        props.webGLView.current.loadProjection(projection)
-    }
+  const onSaveProjectionClick = () => {
+    dispatch(ViewActions.copyFromWorkspace());
+  };
 
-    const onDeleteProjectionClick = (projection: Embedding) => {
-        props.deleteProjection(projection)
-    }
+  const onProjectionClick = (projection: IProjection) => {
+    dispatch(ViewActions.loadById(projection.hash));
+  };
 
+  const onDeleteEditProjectDialog = (handle: string) => {
+    props.deleteProjection(handle);
+    setProjectionToEdit(null);
+  };
 
+  const onCloseEditProjectionDialog = () => {
+    setProjectionToEdit(null);
+  };
 
-    return <div style={{ display: 'flex', flexDirection: 'column', height: '100%' }}>
-        <Box paddingLeft={2} paddingTop={2}>
-            <Typography variant="subtitle2" gutterBottom>{'Projection Methods'}</Typography>
-        </Box>
+  const onSaveEditProjectionDialog = (key: EntityId, changes: any) => {
+    dispatch(ViewActions.save({ id: key, changes }));
+    setProjectionToEdit(null);
+  };
 
-        <Box paddingLeft={2} paddingRight={2}>
-            <Grid container direction="column" spacing={1}>
-                <Grid item>
-                    <Button
-                        style={{
-                            width: '100%'
-                        }}
-                        variant="outlined"
-                        onClick={() => {
-                            setDomainSettings('umap')
-                            setOpen(true)
-                        }}>{'UMAP'}</Button>
-                </Grid>
-                {
-                !frontend_utils.CHEM_PROJECT &&
-                <Grid item>
-                    <Button
-                        style={{
-                            width: '100%'
-                        }}
-                        variant="outlined"
-                        onClick={() => {
-                            setDomainSettings('tsne')
-                            setOpen(true)
-                        }}>{'t-SNE'}</Button>
-                </Grid>
-                }
+  return (
+    <div style={{ display: 'flex', flexDirection: 'column', height: '100%' }}>
+      <EditProjectionDialog
+        projection={projectionToEdit}
+        onClose={onCloseEditProjectionDialog}
+        onSave={onSaveEditProjectionDialog}
+        onDelete={onDeleteEditProjectDialog}
+      />
 
-                {
-                !frontend_utils.CHEM_PROJECT && 
-                <Grid item>
-                    <Button
-                        style={{
-                            width: '100%'
-                        }}
-                        variant="outlined"
-                        onClick={() => {
-                            setDomainSettings('forceatlas2')
-                            setOpen(true)
-                        }}>{'ForceAtlas2'}</Button>
-                </Grid>
-                }
-            </Grid>
-        </Box>
+      <Box paddingLeft={2} paddingTop={2}>
+        <Typography variant="subtitle2" gutterBottom>
+          Projection Methods
+        </Typography>
+      </Box>
 
-        <Box p={1}>
-            <ProjectionControlCard
-                controller={controller}
-                onClose={() => {
-                    if (controller) {
-                        controller.terminate()
-                    }
-                    setController(null)
-                    props.setTrailVisibility(false)
-                }}
-                onComputingChanged={(e, newVal) => {
-                }} />
-        </Box>
+      <Box paddingLeft={2} paddingRight={2}>
+        <EmbeddingMethodButtons setOpen={setOpen} setDomainSettings={setDomainSettings} embeddings={props.config?.embeddings} />
+      </Box>
 
+      <Box p={1}>
+        <ProjectionControlCard
+          dataset_name={props.dataset?.info?.path}
+          controller={controller}
+          onClose={() => {
+            if (controller) {
+              controller.terminate();
+            }
+            setController(null);
+            props.setTrailVisibility(false);
+          }}
+          onComputingChanged={() => {}}
+          onStep={(step) => {
+            // stepRef.current += 1;
+            // setStep(stepRef.current);
+            stepRef.current = step;
+            setStep(step);
+          }}
+        />
+      </Box>
 
-        <GenericSettings
-            projectionParams={props.projectionParams}
-            domainSettings={domainSettings}
-            open={open} onClose={() => setOpen(false)}
-            onStart={(params, selection) => {
-                const checked_sel = selection.filter(s => s.checked)
-                if(checked_sel.length <= 0){
-                    alert("Select at least one feature.")
-                    return;
-                }
+      <GenericSettings
+        // projectionParams={props.projectionParams}
+        domainSettings={domainSettings}
+        open={open}
+        onClose={() => setOpen(false)}
+        onStart={(params, selection) => {
+          const checked_sel = selection.filter((s) => s.checked);
+          if (checked_sel.length <= 0) {
+            alert('Select at least one feature.');
+            return;
+          }
 
-                setOpen(false)
-                props.setProjectionColumns(selection)
-                props.setProjectionParams(params)
+          stepRef.current = 0;
+          setOpen(false);
+          props.setProjectionColumns(selection);
 
-                switch (domainSettings) {
-                    case 'tsne': {
-                        let controller = new TSNEEmbeddingController()
-                        controller.init(props.dataset, selection, params)
-                        controller.stepper = (Y) => {
-                            props.dataset.vectors.forEach((vector, i) => {
-                                vector.x = Y[i][0]
-                                vector.y = Y[i][1]
-                            })
-                            props.webGLView.current.updateXY()
-                            props.webGLView.current.repositionClusters()
-                        }
+          const seed = ADataset.getSpatialData(dataset, workspace.xChannel, workspace.yChannel, workspace.positions);
 
-                        setController(controller)
-                        break;
-                    }
+          // props.setProjectionParams(params)
 
-                    case 'umap': {
-                        let controller = new UMAPEmbeddingController()
-                        let samples = params.useSelection ? props.currentAggregation.aggregation : props.dataset.vectors
+          switch (domainSettings.id) {
+            case ProjectionMethod.TSNE: {
+              const controller = new TSNEEmbeddingController();
+              controller.init(props.dataset, selection, params, seed);
+              const paramsCopy = { ...params };
+              controller.stepper = (Y) => {
+                const workspace = Y.map((y) => ({ x: y[0], y: y[1] }));
+                props.updateWorkspace(workspace, { ...paramsCopy, iterations: stepRef.current, features: [...selection] });
+              };
 
-                        controller.init(props.dataset, selection, params, params.useSelection ? samples : undefined)
-                        controller.stepper = (Y) => {
-                            let source = controller.boundsY(Y)
-                            let target = controller.targetBounds
+              setController(controller);
+              break;
+            }
 
+            case ProjectionMethod.UMAP: {
+              console.log(workspace);
+              const controller = new UMAPEmbeddingController();
+              const paramsCopy = { ...params };
+              controller.init(props.dataset, selection, params, seed);
+              controller.stepper = (Y) => {
+                const workspace = props.dataset.vectors.map((sample, i) => {
+                  return {
+                    x: Y[i][0],
+                    y: Y[i][1],
+                  };
+                });
+                props.updateWorkspace(workspace, { ...paramsCopy, iterations: stepRef.current, features: [...selection] });
+              };
 
+              setController(controller);
+              break;
+            }
+            case ProjectionMethod.FORCEATLAS2: {
+              const controller = new ForceAtlas2EmbeddingController();
+              controller.init(props.dataset, selection, params);
+              const paramsCopy = { ...params };
+              controller.stepper = (Y) => {
+                const workspace = props.dataset.vectors.map((sample) => {
+                  const idx = controller.nodes[sample.__meta__.duplicateOf].__meta__.meshIndex;
+                  return {
+                    x: Y[idx].x,
+                    y: Y[idx].y,
+                  };
+                });
 
-                            samples.forEach((sample, i) => {
-                                if (controller.targetBounds) {
-                                    sample.x = target.x + ((Y[i][0] - source.x) / source.width) * target.width
-                                    sample.y = target.y + ((Y[i][1] - source.y) / source.height) * target.height
-                                } else {
-                                    sample.x = Y[i][0]
-                                    sample.y = Y[i][1]
-                                }
+                props.updateWorkspace(workspace, { ...paramsCopy, iterations: stepRef.current, features: [...selection] });
+              };
 
-                            })
+              setController(controller);
+              break;
+            }
+            default: {
+              // custom embedding controller
+              if (domainSettings.embController) {
+                const controller = domainSettings.embController;
+                const paramsCopy = { ...params };
+                controller.init(props.dataset, selection, params, workspace);
+                controller.stepper = (Y: IBaseProjection) => {
+                  props.updateWorkspace(Y, { ...paramsCopy, method: ProjectionMethod.CUSTOM, iterations: stepRef.current, features: [...selection] });
+                };
 
+                setController(controller);
+              }
+              break;
+            }
+          }
+        }}
+      />
 
-                            props.webGLView.current.updateXY()
-                            props.webGLView.current.repositionClusters()
-                        }
+      <Box paddingLeft={2} paddingTop={2}>
+        <Typography variant="subtitle2" gutterBottom>
+          Projection Settings
+        </Typography>
+      </Box>
 
-                        setController(controller)
-                        break;
-                    }
-                    case 'forceatlas2': {
-                        let controller = new ForceAtlas2EmbeddingController()
-                        controller.init(props.dataset, selection, params)
+      <Box paddingLeft={2} paddingRight={2}>
+        <ClusterTrailSettings />
+      </Box>
 
-                        controller.stepper = (Y) => {
-                            props.dataset.vectors.forEach((sample, i) => {
-                                let idx = controller.nodes[sample.view.duplicateOf].view.meshIndex
-                                sample.x = Y[idx].x
-                                sample.y = Y[idx].y
-                            })
-                            props.webGLView.current.updateXY()
-                        }
-
-                        setController(controller)
-                        break;
-                    }
-
-                }
+      <Box paddingLeft={2} paddingRight={2} display="flex" style={{ flexDirection: 'column', gap: '8px' }}>
+        {dataset && numericFeatures && workspace ? (
+          <SelectFeatureComponent
+            column_info={dataset.columns}
+            label="x channel"
+            default_val={workspace.xChannel}
+            categoryOptions={numericFeatures}
+            onChange={(newValue) => {
+              dispatch(ViewActions.selectChannel({ dataset, channel: 'x', value: newValue }));
             }}
-        ></GenericSettings>
+          />
+        ) : null}
 
-        <Box paddingLeft={2} paddingTop={2}>
-            <Typography variant="subtitle2" gutterBottom>{'Projection Settings'}</Typography>
-        </Box>
+        {dataset && numericFeatures && workspace ? (
+          <SelectFeatureComponent
+            column_info={dataset.columns}
+            label="y channel"
+            default_val={workspace.yChannel}
+            categoryOptions={numericFeatures}
+            onChange={(newValue) => {
+              dispatch(ViewActions.selectChannel({ dataset, channel: 'y', value: newValue }));
+            }}
+          />
+        ) : null}
+      </Box>
 
-        <Box paddingLeft={2} paddingRight={2}>
-            <ClusterTrailSettings></ClusterTrailSettings>
-        </Box>
+      <Box padding={1}>
+        <Typography variant="subtitle2" gutterBottom>
+          Visible Projection
+        </Typography>
+        <ListItem
+          key={workspace?.hash}
+          selected={workspaceIsTemporal === true}
+          secondaryAction={workspaceIsTemporal ? <Chip label="Temporal" variant="outlined" /> : <Chip label="Stored" />}
+        >
+          {workspaceIsTemporal ? (
+            <ListItemText primary={<strong>{workspace?.metadata.method.toUpperCase()}</strong>} secondary={`Iteration ${step}`} />
+          ) : (
+            <ListItemText primary={<strong>{workspace?.metadata.method.toUpperCase()}</strong>} secondary={`loaded from ${workspace?.name}`} />
+          )}
+        </ListItem>
+      </Box>
 
-        <Box paddingLeft={2} paddingTop={2}>
-            <Typography variant="subtitle2" gutterBottom>{'Stored Projections'}</Typography>
-        </Box>
+      <Box paddingLeft={2} paddingRight={2}>
+        <Button onClick={() => onSaveProjectionClick()} variant="outlined" size="small">
+          Store Visible Projection
+        </Button>
+      </Box>
 
-        <Box paddingLeft={2} paddingRight={2}>
-            <Button
-                onClick={(name) => onSaveProjectionClick(new Date().getHours() + ":" + new Date().getMinutes())}
-                variant="outlined"
-                size="small"
-            >{'Store Projection'}</Button>
-        </Box>
+      <Box paddingLeft={2} paddingTop={2}>
+        <Typography variant="subtitle2" gutterBottom>
+          Stored Projections
+        </Typography>
+      </Box>
 
-        <div style={{ overflowY: 'auto', height: '100px', flex: '1 1 auto' }}>
-            <List dense={true}>
-                {props.projections.map(projection => {
-                    return <ListItem key={projection.hash} button onClick={() => onProjectionClick(projection)}>
-                        <ListItemAvatar>
-                            <Avatar>
-                                <FolderIcon />
-                            </Avatar>
-                        </ListItemAvatar>
-                        <ListItemText
-                            primary={`${projection.name}`}
-                            secondary={`${projection.positions.length} items`}
-                        />
-                        <ListItemSecondaryAction>
-                            <IconButton onClick={() => onDeleteProjectionClick(projection)}>
-                                <DeleteIcon />
-                            </IconButton>
-                        </ListItemSecondaryAction>
-                    </ListItem>
-                })}
-            </List>
-        </div>
+      <div style={{ overflowY: 'auto', height: '100px', flex: '1 1 auto' }}>
+        <List dense>
+          {props.projections.ids.map((key) => {
+            const projection = props.projections.entities[key];
+            return (
+              <ListItem
+                key={projection.hash}
+                button
+                onClick={() => onProjectionClick(projection)}
+                selected={workspaceIsTemporal === false && workspace?.hash === projection.hash}
+              >
+                <ListItemText
+                  primary={`${projection.name}`}
+                  secondary={
+                    projection.metadata.method === ProjectionMethod.RANDOM || projection.metadata.method === ProjectionMethod.DATASET
+                      ? 'loaded from dataset'
+                      : `${projection.metadata?.iterations} iterations`
+                  }
+                />
+                <ListItemSecondaryAction>
+                  <IconButton onClick={() => setProjectionToEdit(props.projections.entities[key])}>
+                    <SettingsIcon />
+                  </IconButton>
+                </ListItemSecondaryAction>
+              </ListItem>
+            );
+          })}
+        </List>
+      </div>
     </div>
-})
+  );
+});
