@@ -1,50 +1,59 @@
-import { Box, Button, FormControlLabel, Switch, Typography } from "@material-ui/core";
-import React = require("react");
-import { connect, ConnectedProps } from "react-redux";
-import { setAggregationAction } from "../../Ducks/AggregationDuck";
-import { setHoverWindowMode, WindowMode } from "../../Ducks/HoverSettingsDuck";
-import { SelectionClusters } from "../../Overlays/SelectionClusters/SelectionClusters";
-import { RootState } from "../../Store/Store";
+import { Box, Button, FormControlLabel, Switch, Typography } from '@mui/material';
+import * as React from 'react';
+import { connect, ConnectedProps } from 'react-redux';
+import { selectVectors } from '../../Ducks/AggregationDuck';
+import { setHoverWindowMode, WindowMode } from '../../Ducks/HoverSettingsDuck';
+import { SelectionClusters } from '../../Overlays/SelectionClusters';
+import type { RootState } from '../../Store/Store';
 
 const mapStateToProps = (state: RootState) => ({
-    hoverSettings: state.hoverSettings,
-    currentAggregation: state.currentAggregation,
-    dataset: state.dataset
-})
+  hoverSettings: state.hoverSettings,
+  currentAggregation: state.currentAggregation,
+  dataset: state.dataset,
+  globalLabels: state.globalLabels,
+});
 
-const mapDispatchToProps = dispatch => ({
-    setHoverWindowMode: value => dispatch(setHoverWindowMode(value)),
-    setAggregation: value => dispatch(setAggregationAction(value))
-})
+const mapDispatchToProps = (dispatch) => ({
+  setHoverWindowMode: (value) => dispatch(setHoverWindowMode(value)),
+  setAggregation: (value) => dispatch(selectVectors(value)),
+});
 
 const connector = connect(mapStateToProps, mapDispatchToProps);
 
-type PropsFromRedux = ConnectedProps<typeof connector>
+type PropsFromRedux = ConnectedProps<typeof connector>;
 
-type Props = PropsFromRedux & {
-    hoverUpdate
-}
+type Props = PropsFromRedux & {};
 
+export const HoverTabPanel = connector(({ hoverSettings, setHoverWindowMode, setAggregation, currentAggregation, dataset, globalLabels }: Props) => {
+  const handleChange = (_, value) => {
+    setHoverWindowMode(value ? WindowMode.Extern : WindowMode.Embedded);
+  };
 
-export const HoverTabPanel = connector(({ hoverSettings, setHoverWindowMode, hoverUpdate, setAggregation, currentAggregation, dataset }: Props) => {
-    const handleChange = (_, value) => {
-        setHoverWindowMode(value ? WindowMode.Extern : WindowMode.Embedded)
-    }
+  return (
+    <div style={{ display: 'flex', flexDirection: 'column', height: '100%' }}>
+      <Box paddingLeft={2} paddingTop={2}>
+        {/* TODO: Cluster count not working */}
+        {/* in <b>{currentAggregation.selectedClusters.length}</b> Groups */}
+        <Typography color="textSecondary" variant="body2">
+          Selected <b>{currentAggregation.aggregation.length}</b> out of <b>{dataset && dataset.vectors.length}</b> ${globalLabels.itemLabelPlural}
+        </Typography>
 
-    return <div style={{ display: 'flex', flexDirection: 'column', height: '100%' }}>
-        <Box paddingLeft={2} paddingTop={2}>
-            {/* TODO: Cluster count not working */}
-            {/* in <b>{currentAggregation.selectedClusters.length}</b> Groups */}
-            <Typography color={"textSecondary"} variant="body2">Selected <b>{currentAggregation.aggregation.length}</b> out of <b>{dataset && dataset.vectors.length}</b> items</Typography> 
+        <FormControlLabel
+          control={<Switch color="primary" checked={hoverSettings.windowMode === WindowMode.Extern} onChange={handleChange} name="checkedA" />}
+          label="External Selection Summary"
+        />
 
-            <FormControlLabel
-                control={<Switch color="primary" checked={hoverSettings.windowMode == WindowMode.Extern} onChange={handleChange} name="checkedA" />}
-                label="External Selection Summary"
-            />
+        <Button
+          variant="outlined"
+          onClick={() => {
+            setAggregation([]);
+          }}
+        >
+          Clear Selection
+        </Button>
+      </Box>
 
-            <Button variant="outlined" onClick={() => { setAggregation([]) }}>Clear Selection</Button>
-        </Box>
-
-        <SelectionClusters hoverUpdate={hoverUpdate}></SelectionClusters>
+      <SelectionClusters />
     </div>
-})
+  );
+});
