@@ -1,452 +1,343 @@
-import { connect, ConnectedProps } from 'react-redux'
-import { FunctionComponent } from 'react'
-import * as React from 'react'
-import { Grid, FormControl, InputLabel, Select, MenuItem, Typography, Divider, Box, Accordion, AccordionSummary, AccordionDetails, makeStyles } from '@material-ui/core'
-import { ShapeLegend } from './ShapeLegend/ShapeLegend'
-import { setSelectedVectorByShapeAction } from "../../Ducks/SelectedVectorByShapeDuck"
-import { setVectorByShapeAction } from "../../Ducks/VectorByShapeDuck"
-import { setCheckedShapesAction } from "../../Ducks/CheckedShapesDuck"
-import { RootState } from '../../Store/Store'
-import { setSelectedLineBy } from '../../Ducks/SelectedLineByDuck'
-import { setChannelBrightnessSelection } from '../../Ducks/ChannelBrightnessDuck'
-import { setGlobalPointBrightness } from '../../Ducks/GlobalPointBrightnessDuck'
-import { BrightnessSlider } from './BrightnessSlider/BrightnessSlider'
-import { setChannelSize } from '../../Ducks/ChannelSize'
-import { setGlobalPointSize } from '../../Ducks/GlobalPointSizeDuck'
-import { SizeSlider } from './SizeSlider/SizeSlider'
-import { ColorScaleSelect } from './ColorScaleSelect/ColorScaleSelect'
-import { AdvancedColoringPopover } from './AdvancedColoring/AdvancedColoringPopover/AdvancedColoringPopover'
-import { setChannelColor } from '../../Ducks/ChannelColorDuck'
-import { setAdvancedColoringSelectionAction } from '../../Ducks/AdvancedColoringSelectionDuck'
-import { PathLengthFilter } from './PathLengthFilter/PathLengthFilter'
-import { Legend } from './LineSelection/LineSelection'
-import { LineSelectionTree_GenAlgos, LineSelectionTree_GetChecks, LineTreePopover } from './LineTreePopover/LineTreePopover'
-import { PathBrightnessSlider } from './PathTransparencySlider/PathBrightnessSlider'
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
+/* eslint-disable react-hooks/rules-of-hooks */
+import { connect, ConnectedProps, useDispatch, useSelector } from 'react-redux';
+import * as React from 'react';
+import { Grid, Typography, Box, Accordion, AccordionSummary, AccordionDetails, createFilterOptions, Autocomplete, TextField } from '@mui/material';
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+import { makeStyles } from '@mui/styles';
+import { ShapeLegend } from './ShapeLegend';
+import { setVectorByShapeAction } from '../../Ducks/VectorByShapeDuck';
+import type { RootState } from '../../Store/Store';
+import { setSelectedLineBy } from '../../Ducks/SelectedLineByDuck';
+import { setChannelBrightnessAction } from '../../Ducks/ChannelBrightnessDuck';
+import { setGlobalPointBrightness } from '../../Ducks/GlobalPointBrightnessDuck';
+import { BrightnessSlider } from './BrightnessSlider';
+import { setChannelSize } from '../../Ducks/ChannelSize';
+import { setGlobalPointSize } from '../../Ducks/GlobalPointSizeDuck';
+import { SizeSlider } from './SizeSlider';
+import { ColorScaleSelect } from './ColorScaleSelect';
+import { AdvancedColoringPopover } from './AdvancedColoringPopover';
+import { setChannelColor } from '../../Ducks/ChannelColorDuck';
+import { setAdvancedColoringSelectionAction } from '../../Ducks/AdvancedColoringSelectionDuck';
+import { PathLengthFilter } from './PathLengthFilter';
+import { PathBrightnessSlider } from './PathBrightnessSlider';
+import { CategoryOptionsAPI } from '../../WebGLView/CategoryOptions';
+import { PointDisplayActions } from '../../Ducks/PointDisplayDuck';
+import { BaseColorScale, DefaultFeatureLabel, EncodingChannel } from '../../../model';
+import { ViewSelector } from '../../Ducks/ViewDuck';
+import { ANormalized } from '../../Utility/NormalizedState';
+import { PointColorScaleActions } from '../../Ducks/PointColorScaleDuck';
 
 const mapStateToProps = (state: RootState) => ({
-    selectedVectorByShape: state.selectedVectorByShape,
-    selectedLineBy: state.selectedLineBy,
-    vectorByShape: state.vectorByShape,
-    dataset: state.dataset,
-    categoryOptions: state.categoryOptions,
-    webGlView: state.webGLView,
-    channelBrightness: state.channelBrightness,
-    channelSize: state.channelSize,
-    channelColor: state.channelColor
-})
+  selectedLineBy: state.selectedLineBy,
+  dataset: state.dataset,
+});
 
-const mapDispatchToProps = dispatch => ({
-    setSelectedVectorByShape: selectedVectorByShape => dispatch(setSelectedVectorByShapeAction(selectedVectorByShape)),
-    setVectorByShape: vectorByShape => dispatch(setVectorByShapeAction(vectorByShape)),
-    setCheckedShapes: checkedShapes => dispatch(setCheckedShapesAction(checkedShapes)),
-    setSelectedLineBy: lineBy => dispatch(setSelectedLineBy(lineBy)),
-    setChannelBrightness: value => dispatch(setChannelBrightnessSelection(value)),
-    setGlobalPointBrightness: value => dispatch(setGlobalPointBrightness(value)),
-    setChannelSize: value => dispatch(setChannelSize(value)),
-    setGlobalPointSize: value => dispatch(setGlobalPointSize(value)),
-    setChannelColor: value => dispatch(setChannelColor(value)),
-    setAdvancedColoringSelection: value => dispatch(setAdvancedColoringSelectionAction(value))
-})
+const mapDispatchToProps = (dispatch) => ({
+  setVectorByShape: (vectorByShape) => dispatch(setVectorByShapeAction(vectorByShape)),
+  setSelectedLineBy: (lineBy) => dispatch(setSelectedLineBy(lineBy)),
+  setChannelBrightness: (value) => dispatch(setChannelBrightnessAction(value)),
+  setGlobalPointBrightness: (value) => dispatch(setGlobalPointBrightness(value)),
+  setChannelSize: (value) => dispatch(setChannelSize(value)),
+  setGlobalPointSize: (value) => dispatch(setGlobalPointSize(value)),
+  setAdvancedColoringSelection: (value) => dispatch(setAdvancedColoringSelectionAction(value)),
+});
 
+function SelectFeatureComponent({ label, default_val, categoryOptions, onChange, column_info, datacy }) {
+  let autocomplete_options = [{ value: 'None', inputValue: 'None', group: null }];
+  let autocomplete_filterOptions = null;
+  if (categoryOptions != null) {
+    autocomplete_options = autocomplete_options.concat(
+      categoryOptions.attributes.map((attribute) => {
+        let group = DefaultFeatureLabel;
+        if (column_info != null && attribute.key in column_info) {
+          group = column_info[attribute.key].featureLabel;
+        }
+        return {
+          value: attribute.key,
+          inputValue: attribute.name,
+          group,
+        };
+      }),
+    );
+    autocomplete_filterOptions = createFilterOptions({
+      stringify: (option: any) => {
+        return option.group + option.value;
+      },
+    });
+  }
 
+  return (
+    <Autocomplete
+      id={`vectorBySelect_${label}`}
+      data-cy={datacy}
+      filterOptions={autocomplete_filterOptions}
+      onChange={(event, newValue) => {
+        // @ts-ignore
+        if (newValue) onChange(newValue.value);
+      }}
+      options={autocomplete_options.sort((a, b) => {
+        if (a.value === 'None') return -1;
+        if (b.value === 'None') return 1;
 
+        if (a.group === b.group) {
+          return -b.value.localeCompare(a.value);
+        }
+        return -b.group.localeCompare(a.group);
+      })}
+      size="small"
+      groupBy={(option: any) => option.group}
+      getOptionLabel={(option: any) => option.inputValue}
+      isOptionEqualToValue={(option: any, value) => {
+        return option.value === value.value;
+      }}
+      // @ts-ignore
+      value={default_val ? autocomplete_options.filter((option: any) => option.value === default_val.key)[0] : autocomplete_options[0]}
+      renderInput={(params) => <TextField {...params} label={`${label} by`} />}
+    />
+  );
+}
 
 const connector = connect(mapStateToProps, mapDispatchToProps, null, { forwardRef: true });
 
-type PropsFromRedux = ConnectedProps<typeof connector>
+type PropsFromRedux = ConnectedProps<typeof connector>;
 
 type Props = PropsFromRedux & {
-    lineColorScheme
-}
+  encodings: EncodingChannel[];
+};
 
-/**
- 
-
-        {
-            <FormControl style={{ margin: '4px 0px' }}>
-                <InputLabel shrink id="lineByLabel">{"line by"}</InputLabel>
-                <Select labelId="lineByLabel"
-                    id="lineBySelect"
-                    displayEmpty
-                    value={selectedLineBy.value}
-                    onChange={(event) => {
-                        setSelectedLineBy(event.target.value)
-                        webGlView.current.recreateLines(event.target.value)
-                    }}
-                >
-                    <MenuItem value="">None</MenuItem>
-                    {
-                        selectedLineBy.options.map((option, i) => {
-                            return <MenuItem key={option} value={option}>{option}</MenuItem>
-                        })
-                    }
-                </Select>
-            </FormControl>
-        }
-
- */
-
-const useStyles = makeStyles((theme) => ({
-    root: {
-        width: '100%',
-    },
-    heading: {
-        fontSize: theme.typography.pxToRem(15),
-        flexBasis: '33.33%',
-        flexShrink: 0,
-    },
-    secondaryHeading: {
-        fontSize: theme.typography.pxToRem(15),
-        color: theme.palette.text.secondary,
-    },
-    details: {
-        padding: '0px',
-        display: 'flex',
-        flexDirection: 'column'
-    }
+const useStyles = makeStyles(() => ({
+  root: {
+    width: '100%',
+  },
+  heading: {
+    // fontSize: theme.typography.pxToRem(15),
+    flexBasis: '33.33%',
+    flexShrink: 0,
+  },
+  secondaryHeading: {
+    // fontSize: theme.typography.pxToRem(15),
+    // color: theme.palette.text.secondary,
+  },
+  details: {
+    padding: '0px',
+    display: 'flex',
+    flexDirection: 'column',
+  },
 }));
 
+export function StatesTabPanelFull({
+  dataset,
+  setVectorByShape,
+  setChannelBrightness,
+  setGlobalPointBrightness,
+  setChannelSize,
+  setGlobalPointSize,
+  encodings,
+  setAdvancedColoringSelection,
+}: Props) {
+  if (dataset == null) {
+    return null;
+  }
 
-export const StatesTabPanelFull = ({
-    selectedVectorByShape,
+  const activeMultiple = useSelector(ViewSelector.defaultSelector);
+  const {
     vectorByShape,
-    dataset,
-    setSelectedVectorByShape,
-    setVectorByShape,
-    setCheckedShapes,
-    categoryOptions,
-    selectedLineBy,
-    setSelectedLineBy,
-    webGlView,
     channelBrightness,
-    setChannelBrightness,
-    setGlobalPointBrightness,
     channelSize,
-    setChannelSize,
-    setGlobalPointSize,
     channelColor,
-    setChannelColor,
-    setAdvancedColoringSelection,
-    lineColorScheme
-}: Props) => {
-    if (dataset == null) {
-        return null;
-    }
+    pointColorMapping,
+    pointColorScale,
+    lineBrightness,
+    pathLengthRange,
+    globalPointSize,
+    globalPointBrightness,
+  } = activeMultiple.attributes;
+  const active = useSelector<RootState, BaseColorScale>((state) => ANormalized.get(state.colorScales.scales, pointColorScale as string));
 
-    const classes = useStyles();
+  const dispatch = useDispatch();
 
+  const classes = useStyles();
 
+  const [expanded, setExpanded] = React.useState<boolean | string>(false);
 
+  const handleChange = (panel) => (event, isExpanded) => {
+    setExpanded(isExpanded ? panel : false);
+  };
 
+  const categoryOptions = dataset?.categories;
 
+  const points_box = (
+    <Box
+      sx={{
+        '& .MuiAutocomplete-root': { p: 2, width: '100%', boxSizing: 'border-box' },
+      }}
+    >
+      {(!encodings || encodings.includes(EncodingChannel.Shape)) && categoryOptions != null && CategoryOptionsAPI.hasCategory(categoryOptions, 'shape') ? (
+        <SelectFeatureComponent
+          column_info={dataset?.columns}
+          datacy="shape-encoding-select"
+          label="shape"
+          default_val={vectorByShape}
+          categoryOptions={CategoryOptionsAPI.getCategory(categoryOptions, 'shape')}
+          onChange={(newValue) => {
+            let attribute = CategoryOptionsAPI.getCategory(categoryOptions, 'shape').attributes.filter((a) => a.key === newValue)[0];
 
+            if (attribute === undefined) {
+              attribute = null;
+            }
+            setVectorByShape(attribute);
+          }}
+        />
+      ) : (
+        <div />
+      )}
 
-
-    const [expanded, setExpanded] = React.useState<boolean | string>(false);
-
-    const handleChange = (panel) => (event, isExpanded) => {
-        setExpanded(isExpanded ? panel : false);
-    };
-
-    const points_box = <Box>
-        {
-            categoryOptions != null && categoryOptions.hasCategory("shape") ?
-                <Grid
-                    container
-                    justify="center"
-                    alignItems="stretch"
-                    direction="column"
-                    style={{ padding: '0 16px' }}>
-                    <FormControl style={{ margin: '4px 0px' }}>
-                        <InputLabel shrink id="vectorByShapeSelectLabel">{"shape by"}</InputLabel>
-                        <Select labelId="vectorByShapeSelectLabel"
-                            id="vectorByShapeSelect"
-                            displayEmpty
-                            value={selectedVectorByShape}
-                            onChange={(event) => {
-                                setSelectedVectorByShape(event.target.value)
-
-                                if (event.target.value != null && event.target.value != "") {
-                                    var attribute = categoryOptions.getCategory("shape").attributes.filter(a => a.key == event.target.value)[0]
-                                    setVectorByShape(attribute)
-                                } else {
-                                    setVectorByShape(null)
-                                }
-                            }}
-                        >
-                            <MenuItem value="">None</MenuItem>
-                            {categoryOptions.getCategory("shape").attributes.map(attribute => {
-                                return <MenuItem key={attribute.key} value={attribute.key}>{attribute.name}</MenuItem>
-                            })}
-                        </Select>
-                    </FormControl>
-                </Grid>
-
-                :
-                <div></div>
-        }
-
+      {(!encodings || encodings.includes(EncodingChannel.Shape)) && vectorByShape && (
         <Grid item style={{ padding: '0 16px' }}>
-            <ShapeLegend
-                dataset={dataset}
-                category={vectorByShape}
-                onChange={(checkboxes) => {
-                    setCheckedShapes(checkboxes)
-                }}></ShapeLegend>
+          <ShapeLegend
+            dataset={dataset}
+            category={vectorByShape}
+            onChange={(checkboxes) => {
+              dispatch(PointDisplayActions.setCheckedShapes(checkboxes));
+            }}
+          />
         </Grid>
+      )}
 
+      {categoryOptions != null && CategoryOptionsAPI.hasCategory(categoryOptions, 'transparency') ? (
+        <SelectFeatureComponent
+          column_info={dataset?.columns}
+          datacy="brightness-encoding-select"
+          label="brightness"
+          default_val={channelBrightness}
+          categoryOptions={CategoryOptionsAPI.getCategory(categoryOptions, 'transparency')}
+          onChange={(newValue) => {
+            let attribute = CategoryOptionsAPI.getCategory(categoryOptions, 'transparency').attributes.filter((a) => a.key === newValue)[0];
 
-        {
-            categoryOptions != null && categoryOptions.hasCategory("transparency") ?
-                <Grid
-                    container
-                    justify="center"
-                    alignItems="stretch"
-                    direction="column"
-                    style={{ padding: '0 16px' }}>
-                    <FormControl style={{ margin: '4px 0px' }}>
-                        <InputLabel shrink id="vectorByTransparencySelectLabel">{"brightness by"}</InputLabel>
-                        <Select labelId="vectorByTransparencySelectLabel"
-                            id="vectorByTransparencySelect"
-                            displayEmpty
-                            value={channelBrightness ? channelBrightness.key : ''}
-                            onChange={(event) => {
-                                var attribute = categoryOptions.getCategory("transparency").attributes.filter(a => a.key == event.target.value)[0]
-
-                                if (attribute == undefined) {
-                                    attribute = null
-                                }
-
-                                let pointBrightness = attribute ? [0.25, 1] : [1]
-
-                                setGlobalPointBrightness(pointBrightness)
-                                setChannelBrightness(attribute)
-                                webGlView.current.particles.transparencyCat(attribute, pointBrightness)
-                                webGlView.current.requestRender()
-                            }}
-                        >
-                            <MenuItem value="">None</MenuItem>
-                            {categoryOptions.getCategory("transparency").attributes.map(attribute => {
-                                return <MenuItem key={attribute.key} value={attribute.key}>{attribute.name}</MenuItem>
-                            })}
-                        </Select>
-                    </FormControl>
-                </Grid>
-                :
-                <div></div>
-        }
-
-        <BrightnessSlider></BrightnessSlider>
-
-
-
-
-        {
-            categoryOptions != null && categoryOptions.hasCategory("size") ?
-                <Grid
-                    container
-                    justify="center"
-                    alignItems="stretch"
-                    direction="column"
-                    style={{ padding: '0 16px' }}>
-                    <FormControl style={{ margin: '4px 0px' }}>
-                        <InputLabel shrink id="vectorBySizeSelectLabel">{"size by"}</InputLabel>
-                        <Select labelId="vectorBySizeSelectLabel"
-                            id="vectorBySizeSelect"
-                            displayEmpty
-                            value={channelSize ? channelSize.key : ''}
-                            onChange={(event) => {
-                                var attribute = categoryOptions.getCategory("size").attributes.filter(a => a.key == event.target.value)[0]
-                                if (attribute == undefined) {
-                                    attribute = null
-                                }
-
-                                let pointSize = attribute ? [1, 2] : [1]
-
-                                setGlobalPointSize(pointSize)
-
-                                setChannelSize(attribute)
-
-                                webGlView.current.particles.sizeCat(attribute, pointSize)
-                            }}
-                        >
-                            <MenuItem value="">None</MenuItem>
-                            {categoryOptions.getCategory("size").attributes.map(attribute => {
-                                return <MenuItem key={attribute.key} value={attribute.key}>{attribute.name}</MenuItem>
-                            })}
-                        </Select>
-                    </FormControl>
-                </Grid>
-                :
-                <div></div>
-        }
-
-        <SizeSlider></SizeSlider>
-
-
-        {
-            categoryOptions != null && categoryOptions.hasCategory("color") ?
-                <Grid
-                    container
-                    item
-                    alignItems="stretch"
-                    direction="column"
-                    style={{ padding: '0 16px' }}
-                >
-
-                    <Grid container item alignItems="stretch" direction="column">
-                        <FormControl style={{ margin: '4px 0px' }}>
-                            <InputLabel shrink id="vectorByColorSelectLabel">{"color by"}</InputLabel>
-                            <Select labelId="vectorByColorSelectLabel"
-                                id="vectorByColorSelect"
-                                displayEmpty
-                                value={channelColor ? channelColor.key : ""}
-                                onChange={(event) => {
-                                    var attribute = null
-                                    if (event.target.value != "") {
-                                        attribute = categoryOptions.getCategory("color").attributes.filter(a => a.key == event.target.value)[0]
-                                    }
-
-                                    setAdvancedColoringSelection(new Array(10000).fill(true))
-                                    setChannelColor(attribute)
-                                }}
-                            >
-                                <MenuItem value="">None</MenuItem>
-                                {categoryOptions.getCategory("color").attributes.map(attribute => {
-                                    return <MenuItem key={attribute.key} value={attribute.key}>{attribute.name}</MenuItem>
-                                })}
-                            </Select>
-                        </FormControl>
-                    </Grid>
-                </Grid>
-                :
-                <div></div>
-        }
-
-        <Grid item>
-
-            <ColorScaleSelect></ColorScaleSelect>
-        </Grid>
-
-
-        <Grid item style={{ padding: '16px 0px' }}>
-            {
-                channelColor != null && channelColor.type == 'categorical' ?
-
-                    <AdvancedColoringPopover></AdvancedColoringPopover>
-                    :
-                    <div></div>
+            if (attribute === undefined) {
+              attribute = null;
             }
 
+            const pointBrightness = attribute ? [0.25, 1] : [1];
+            setGlobalPointBrightness(pointBrightness);
+            setChannelBrightness(attribute);
+          }}
+        />
+      ) : (
+        <div />
+      )}
 
+      <BrightnessSlider globalPointBrightness={globalPointBrightness} />
+
+      {categoryOptions != null && CategoryOptionsAPI.hasCategory(categoryOptions, 'size') ? (
+        <SelectFeatureComponent
+          column_info={dataset?.columns}
+          datacy="size-encoding-select"
+          label="size"
+          default_val={channelSize}
+          categoryOptions={CategoryOptionsAPI.getCategory(categoryOptions, 'size')}
+          onChange={(newValue) => {
+            let attribute = CategoryOptionsAPI.getCategory(categoryOptions, 'size').attributes.filter((a) => a.key === newValue)[0];
+            if (attribute === undefined) {
+              attribute = null;
+            }
+
+            const pointSize = attribute ? [1, 2] : [1];
+
+            setGlobalPointSize(pointSize);
+
+            setChannelSize(attribute);
+          }}
+        />
+      ) : (
+        <div />
+      )}
+
+      <SizeSlider globalPointSize={globalPointSize} />
+
+      {categoryOptions != null && CategoryOptionsAPI.hasCategory(categoryOptions, 'color') ? (
+        <SelectFeatureComponent
+          column_info={dataset?.columns}
+          datacy="color-encoding-select"
+          label="color"
+          default_val={channelColor}
+          categoryOptions={CategoryOptionsAPI.getCategory(categoryOptions, 'color')}
+          onChange={(newValue) => {
+            let attribute = null;
+            if (newValue && newValue !== '') {
+              attribute = CategoryOptionsAPI.getCategory(categoryOptions, 'color').attributes.filter((a) => a.key === newValue)[0];
+            }
+            if (attribute === undefined) {
+              attribute = null;
+            }
+
+            setAdvancedColoringSelection(new Array(10000).fill(true));
+            dispatch(setChannelColor(attribute));
+
+            if (attribute) {
+              dispatch(PointColorScaleActions.initScaleByType(attribute.type));
+            }
+          }}
+        />
+      ) : (
+        <div />
+      )}
+
+      <Grid item>
+        <ColorScaleSelect channelColor={channelColor} active={active} />
+      </Grid>
+
+      {channelColor != null ? (
+        <Grid item style={{ padding: '16px 0px' }}>
+          {channelColor.type === 'categorical' ? <AdvancedColoringPopover pointColorMapping={pointColorMapping} /> : null}
         </Grid>
+      ) : null}
     </Box>
+  );
 
-    const accordion = <div style={{
-    }}>
-        <Accordion expanded={expanded === 'panel1'} onChange={handleChange('panel1')}>
-            <AccordionSummary
-                expandIcon={<ExpandMoreIcon />}
-                aria-controls="panel1bh-content"
-                id="panel1bh-header"
-            >
-                <Typography className={classes.heading}>Lines</Typography>
-            </AccordionSummary>
-            <AccordionDetails className={classes.details}>
-                {dataset && dataset.isSequential && <div>
-                    <Grid
-                        container
-                        justify="center"
-                        alignItems="stretch"
-                        direction="column"
-                        style={{ padding: '0 16px' }}>
-                        {//<Legend
-                            //   ref={this.legend}
-                            //   onLineSelect={this.onLineSelect}></Legend>
-
-                        }
-
-
-                        <Box p={1}></Box>
-
-                        <LineTreePopover
-                            webGlView={webGlView}
-                            dataset={dataset}
-                            colorScale={lineColorScheme} />
-                    </Grid>
-
-
-                    <div style={{ margin: '8px 0px' }}></div>
-
-                    <PathLengthFilter></PathLengthFilter>
-                    <PathBrightnessSlider></PathBrightnessSlider>
-                </div>
+  const accordion = (
+    <div style={{}}>
+      <Accordion expanded={expanded === 'panel1'} onChange={handleChange('panel1')}>
+        <AccordionSummary expandIcon={<ExpandMoreIcon />} aria-controls="panel1bh-content" id="panel1bh-header">
+          <Typography className={classes.heading}>Lines</Typography>
+        </AccordionSummary>
+        <AccordionDetails className={classes.details}>
+          {dataset && dataset.isSequential && (
+            <div>
+              <Grid container justifyContent="center" alignItems="stretch" direction="column" style={{ padding: '0 16px' }}>
+                {
+                  // <Legend
+                  //   ref={this.legend}
+                  //   onLineSelect={this.onLineSelect}></Legend>
                 }
-            </AccordionDetails>
-        </Accordion>
 
+                <Box p={1} />
 
-        <Accordion expanded={expanded === 'panel2'} onChange={handleChange('panel2')}>
-            <AccordionSummary
-                expandIcon={<ExpandMoreIcon />}
-                aria-controls="panel1bh-content"
-                id="panel1bh-header"
-            >
-                <Typography className={classes.heading}>Points</Typography>
-            </AccordionSummary>
-            <AccordionDetails className={classes.details}>
-                {points_box}
-            </AccordionDetails>
-        </Accordion>
-    </div>;
+                {/** <LineTreePopover
+                            webGlView={webGLView}
+                            dataset={dataset}
+                        colorScale={lineColorScheme} />* */}
+              </Grid>
 
-    return <div>{dataset && dataset.isSequential ? accordion : points_box}</div>
+              <div style={{ margin: '8px 0px' }} />
+
+              <PathLengthFilter pathLengthRange={pathLengthRange} />
+              <PathBrightnessSlider lineBrightness={lineBrightness} />
+            </div>
+          )}
+        </AccordionDetails>
+      </Accordion>
+
+      <Accordion expanded={expanded === 'panel2'} onChange={handleChange('panel2')}>
+        <AccordionSummary expandIcon={<ExpandMoreIcon />} aria-controls="panel1bh-content" id="panel1bh-header">
+          <Typography className={classes.heading}>Points</Typography>
+        </AccordionSummary>
+        <AccordionDetails className={classes.details}>{points_box}</AccordionDetails>
+      </Accordion>
+    </div>
+  );
+
+  return <div>{dataset && dataset.isSequential ? accordion : points_box}</div>;
 }
 
-
-export const StatesTabPanel = connector(StatesTabPanelFull)
-
-
-/**
- *
-         {
-            categoryOptions != null && categoryOptions.hasCategory("size") ?
-                <Grid
-                    container
-                    justify="center"
-                    alignItems="stretch"
-                    direction="column"
-                    style={{ padding: '0 16px' }}>
-                    <FormControl style={{ margin: '4px 0px' }}>
-                        <InputLabel shrink id="vectorBySizeSelectLabel">{"size by"}</InputLabel>
-                        <Select labelId="vectorBySizeSelectLabel"
-                            id="vectorBySizeSelect"
-                            displayEmpty
-                            value={channelSize ? channelSize.key : ''}
-                            onChange={(event) => {
-                                var attribute = categoryOptions.getCategory("size").attributes.filter(a => a.key == event.target.value)[0]
-                                if (attribute == undefined) {
-                                    attribute = null
-                                }
-
-                                let pointSize = attribute ? [1, 2] : [1]
-
-                                setGlobalPointSize(pointSize)
-
-                                setChannelSize(attribute)
-
-                                webGlView.current.particles.sizeCat(attribute, pointSize)
-                            }}
-                        >
-                            <MenuItem value="">None</MenuItem>
-                            {categoryOptions.getCategory("size").attributes.map(attribute => {
-                                return <MenuItem key={attribute.key} value={attribute.key}>{attribute.name}</MenuItem>
-                            })}
-                        </Select>
-                    </FormControl>
-                </Grid>
-                :
-                <div></div>
-        }
- */
+export const StatesTabPanel = connector(StatesTabPanelFull);
