@@ -1,18 +1,10 @@
-const SET = "ducks/channelColor/SET"
+import { createAction, createReducer } from '@reduxjs/toolkit';
+import { CategoryOption } from '../WebGLView/CategoryOptions';
 
-const channelColor = (state = null, action) => {
-    switch (action.type) {
-        case SET:
-            return action.channelColor
-        default:
-            return state
-    }
-}
+export const setChannelColor = createAction<CategoryOption>('ducks/channelColor/SET');
 
-
-export const setChannelColor = channelColor => ({
-    type: SET,
-    channelColor: channelColor
-})
-
-export default channelColor
+export const channelColor = createReducer<CategoryOption>(null, (builder) => {
+  builder.addCase(setChannelColor, (_, action) => {
+    return action.payload;
+  });
+});
