@@ -1,33 +1,30 @@
-import Cluster from "../Utility/Data/Cluster";
-import { Vect } from "../Utility/Data/Vect";
+import { ICluster } from '../../model/ICluster';
+import { IVector } from '../../model/Vector';
 
-const SET = "ducks/hoverState/SET"
+const SET = 'ducks/hoverState/SET';
 
 export const setHoverState = (hoverState, updater) => ({
-    type: SET,
-    input: {data: hoverState, updater: updater}
+  type: SET,
+  input: { data: hoverState, updater },
 });
 
-
 const initialState: HoverStateType = {
-    data: null,
-    updater: ""
-}
+  data: null,
+  updater: '',
+};
 
 export type HoverStateType = {
-    data: Vect | Cluster,
-    updater: String
-}
-
+  data: IVector | ICluster;
+  updater: string;
+};
 
 const hoverState = (state = initialState, action): HoverStateType => {
-    switch (action.type) {
-        case SET:
-            return {...state, data: action.input?.data, updater: action.input?.updater}
-        default:
-            return state
-    }
-}
+  switch (action.type) {
+    case SET:
+      return { ...state, data: action.input?.data, updater: action.input?.updater };
+    default:
+      return state;
+  }
+};
 
-export default hoverState
-
+export default hoverState;
