@@ -1,14 +1,6 @@
-import { QualitativeScaleMapping } from "./QualitativeScaleMapping";
-
-
-
-
-
-
-
-
 export class ColorScheme {
   colors: any;
+
   mapping: any;
 
   constructor(colors) {
@@ -17,18 +9,6 @@ export class ColorScheme {
 
   getMapping() {
     return this.mapping;
-  }
-
-  createMapping(values) {
-
-    var i = 0;
-    this.mapping = values.reduce((map, obj) => {
-      map[obj] = this.colors[i % this.colors.length];
-      i++;
-      return map;
-    }, {});
-
-    return new QualitativeScaleMapping(this, values);
   }
 
   map(value) {
