@@ -1,6 +1,4 @@
 export interface Loader {
-    resolvePath(entry, finished)
-    resolveContent(content, finished)
-
-    
+  resolvePath(entry, finished);
+  resolveContent(content, finished);
 }
